Fix clockTick build
Only override clockTick on Darwin/Linux.
